requirements.txt
instabot>=0.117.0
Pillow>=7.2.0
requests>=2.24.0
urllib3>=1.25.10